Tool improvements

+++ build.py
@@ -21,6 +21,17 @@ class statement_metadata:
         self.timelimit = timelimit
         self.author = author
 
+def create_parser():
+    parser = argparse.ArgumentParser(formatter_class=argparse.RawTextHelpFormatter)
+    parser.add_argument('-a', '--all', action='store_true',
+                        default=False, help='apply action on all problems')
+    parser.add_argument('-i', '--interactive', action='store_true',
+                        default=False, help='set problem to interative on init')
+    parser.add_argument(
+        'mode', choices=['init', 'build', 'genio', 'genpdf', 'pack2boca'], help='init: init a problem.\nbuild: build a problem.\ngenio: gen problem input/output while validating inputs.\ngenpdf: generates problem and tutorial PDFs.\npack2boca: pack a problem to BOCA format.\n')
+    parser.add_argument('problem_id', nargs='?')
+    return parser
+
 
 def genio(problem_folder):
     build_executables(problem_folder)
@@ -37,14 +48,10 @@ def build(problem_folder):
     genpdf(problem_folder)
 
 
-def build_all():
-    problem_ids = [f for f in os.listdir(
-        'Problemas') if os.path.isdir(os.path.join('Problemas', f))]
-    for id in problem_ids:
-        build(id)
-
-
 def init(problem_folder, interactive=False):
+    if (os.path.exists(problem_folder)):
+        print("Problem ID already exists in the directory")
+        sys.exit(1)
     folder = os.path.join(os.path.dirname(
         os.path.abspath(__file__)), 'arquivos')
     shutil.copytree(folder, problem_folder,
@@ -69,13 +76,6 @@ def pack2boca(problem_folder):
     boca_pack(problem_folder)
 
 
-def packall2boca():
-    problem_ids = [f for f in os.listdir(
-        'Problemas') if os.path.isdir(os.path.join('Problemas', f))]
-    for id in problem_ids:
-        pack2boca(id)
-
-
 def pack2uri(problem_id):
     print('Not implemented')
     pass
@@ -87,14 +87,7 @@ def packall2uri():
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-a', '--all', action='store_true',
-                        default=False, help='apply action on all problems')
-    parser.add_argument('-i', '--interactive', action='store_true',
-                        default=False, help='set problem do interative on init')
-    parser.add_argument(
-        'mode', choices=['init', 'build', 'buildall', 'genio', 'genpdf', 'pack2boca', 'packall2boca'], help='\ninit: init a problem\nbuild: build  a problem.\ngenio: generates problem input/output while validating inputs\ngenpdf: generates problem and tutorial PDFs.\npack2boca: pack a problem to BOCA format.\n')
-    parser.add_argument('problem_id', nargs='?')
+    parser = create_parser()
     args = parser.parse_args()
     if(not args.all and not args.problem_id):
         parser.error(args.mode + ' mode requires a problem id. Usage:' +
@@ -104,16 +97,10 @@ if __name__ == "__main__":
         init(args.problem_id, args.interactive)
         print('Problem', args.problem_id, 'initialized')
     elif(args.mode == 'build'):
-        if(not args.all):
-            print("Building problem", args.problem_id)
-            build(args.problem_id)
-        else:
-            build_all()
+        print("Building problem", args.problem_id)
+        build(args.problem_id)
     elif(args.mode == 'pack2boca'):
-        if(not args.all):
-            pack2boca(args.problem_id)
-        else:
-            packall2boca()
+        pack2boca(args.problem_id)
     elif(args.mode == 'genpdf'):
         genpdf(args.problem_id)
     elif(args.mode == 'genio'):
